forge fmt

+++ contracts/test/BoundlessMarket.t.sol
@@ -384,12 +384,7 @@ contract BoundlessMarketTest is Test {
             PredicateType predicateType = requests[i].requirements.predicate.predicateType;
             if (predicateType != PredicateType.ClaimDigestMatch) {
                 claimDigest = ReceiptClaimLib.ok(requests[i].requirements.imageId, sha256(journal)).digest();
-                callbackData = abi.encode(
-                    CallbackData({
-                        imageId: requests[i].requirements.imageId,
-                        journal: journal
-                    })
-                );
+                callbackData = abi.encode(CallbackData({imageId: requests[i].requirements.imageId, journal: journal}));
             } else {
                 claimDigest = bytesToBytes32(requests[i].requirements.predicate.data);
             }
@@ -502,7 +497,7 @@ contract BoundlessMarketTest is Test {
 
     function bytesToBytes32(bytes memory b) internal pure returns (bytes32) {
         bytes32 out;
-        for (uint i = 0; i < 32; i++) {
+        for (uint256 i = 0; i < 32; i++) {
             out |= bytes32(b[i] & 0xFF) >> (i * 8);
         }
         return out;
